Detecting login provider
As mentioned in previous step, now app is detecting login provider of current user. Depending on this provider user will be able to change email and password or not. If provider isn't an email-password method, there will be displayed info about login provider of the user instead of change email/password options.